Switching to working with games with a threshold of concurrent daily users

+++ scripts/download_images.py
@@ -6,21 +6,10 @@ from pathlib import Path
 
 import pandas as pd
 
-IMG_DIR = Path("data/img")
-os.makedirs(IMG_DIR, exist_ok=True)
-
-HEADER_DIR = Path(f"{IMG_DIR}/header")
-os.makedirs(HEADER_DIR, exist_ok=True)
-
-# CSV file path
-CSV_FILE = "top_genre_headerlinks.csv"
-
-def create_csv(top_n):
-    data_dir = "data/oct24_clean/"
-
+def get_top_concurrent_df(data_dir, top_n):
     games = pd.read_csv(data_dir + "games.csv")
     genres = pd.read_csv(data_dir + "genres_translated.csv")
-    spy = pd.read_csv(data_dir + "steamspy_insights.csv") 
+    spy = pd.read_csv(data_dir + "steamspy_insights.csv")
     prom = pd.read_csv(data_dir + "promotional.csv")
 
     # Get player numbers
@@ -44,6 +33,32 @@ def create_csv(top_n):
     df_download = df_download.loc[:, ["app_id", "name", "header_image"]]
     df_download = df_download.drop_duplicates(subset="app_id")
 
+    return df_download
+
+def get_ge_n_users_df(data_dir, data_fname):
+    # games = pd.read_csv(data_dir + "games.csv")
+    prom = pd.read_csv(data_dir + "promotional.csv")
+
+    df = pd.read_csv(f"{data_dir}/{data_fname}")
+
+    # df = games.merge(df, on="app_id", how="right")
+
+    df_download = df.merge(prom.loc[:, ["app_id", "header_image"]], how="left", on="app_id")
+    df_download = df_download.loc[:, ["app_id", "header_image"]]
+    df_download = df_download.drop_duplicates(subset="app_id")
+
+    return df_download
+
+def create_csv():
+    data_dir = "data/oct24_clean/"
+    data_fname = "ge-10_concurrent-users.csv"
+
+    # df_download = get_top_concurrent_df(data_dir, top_n)
+    df_download = get_ge_n_users_df(data_dir, data_fname)
+
+    print(df_download.info())
+    print(df_download.tail(10))
+
     df_download.to_csv(IMG_DIR / CSV_FILE, index=False)
 
 def sanitize_filename(name):
@@ -65,6 +80,11 @@ def download_images():
             if response.status_code == 200:
                 # Save image
                 image_path = os.path.join(HEADER_DIR, f"{app_id}.jpg")
+
+                if os.path.exists(image_path):
+                    print(f"{image_path} exists, skipping...")
+                    continue
+
                 with open(image_path, "wb") as img_file:
                     for chunk in response.iter_content(1024):
                         img_file.write(chunk)
@@ -73,6 +93,15 @@ def download_images():
                 print(f"Failed to download: {image_url}")
 
 if __name__ == "__main__":
-    create_csv(top_n=100)
+    IMG_DIR = Path("data/img")
+    os.makedirs(IMG_DIR, exist_ok=True)
+
+    HEADER_DIR = Path(f"{IMG_DIR}/header")
+    os.makedirs(HEADER_DIR, exist_ok=True)
+
+    # CSV file path
+    CSV_FILE = "ge-10_concurrent-users_header-links.csv"
+
+    create_csv()
     download_images()
